# Improve performance by for pid stats (cgroups1) re-using readuint